Previous time added

+++ app/src/main/java/com/example/pendulumtestjava/fragments/pendulumsFragments/settings/DoublePendulumSettings.java
@@ -107,9 +107,10 @@ public class DoublePendulumSettings extends AppCompatDialogFragment {
                             data.setA2(Double.parseDouble(a2.getText().toString()));
                             data.setR1(Double.parseDouble(r1.getText().toString()));
                             data.setR2(Double.parseDouble(r2.getText().toString()));
-                            data.setG((float) Double.parseDouble(g.getText().toString()));
+                            data.setG(Float.parseFloat(g.getText().toString()));
                             data.setM1(Double.parseDouble(m1.getText().toString()));
                             data.setM2(Double.parseDouble(m2.getText().toString()));
+
                             data.setBall1Color(ball1DefaultColor);
                             data.setBall2Color(ball2DefaultColor);
                             data.setTrace1Color(trace1DefaultColor);
